Add cmake option `ENABLE_TIMER_LOOP`.
We use `ENABLE_TIMER_LOOP` to decide if we should enable a timer-based
function for our main loop or keep using wxIdleEvents triggers.

The timer was meant to replace wxIdleEvents only on Windows, since we
have a couple of issues related. On Linux, it should be **VERY** buggy
as of current state.

The default is: enabled on Windows only.